Move `WorkloadIdentitySelector` to bot package
