+++ lib/tbot/config/config_test.go
@@ -295,7 +295,7 @@ func TestBotConfig_YAML(t *testing.T) {
 						Destination: &destination.Directory{
 							Path: "/an/output/path",
 						},
-						Selector: WorkloadIdentitySelector{
+						Selector: bot.WorkloadIdentitySelector{
 							Name: "my-workload-identity",
 						},
 						CredentialLifetime: bot.CredentialLifetime{
@@ -305,7 +305,7 @@ func TestBotConfig_YAML(t *testing.T) {
 					},
 					&WorkloadIdentityAPIService{
 						Listen: "tcp://127.0.0.1:123",
-						Selector: WorkloadIdentitySelector{
+						Selector: bot.WorkloadIdentitySelector{
 							Name: "my-workload-identity",
 						},
 						CredentialLifetime: bot.CredentialLifetime{
@@ -317,7 +317,7 @@ func TestBotConfig_YAML(t *testing.T) {
 						Destination: &destination.Directory{
 							Path: "/an/output/path",
 						},
-						Selector: WorkloadIdentitySelector{
+						Selector: bot.WorkloadIdentitySelector{
 							Name: "my-workload-identity",
 						},
 						Audiences: []string{"audience1", "audience2"},
