+++ lib/tbot/config/service_workload_identity_api_test.go
@@ -44,7 +44,7 @@ func TestWorkloadIdentityAPIService_YAML(t *testing.T) {
 						},
 					},
 				},
-				Selector: WorkloadIdentitySelector{
+				Selector: bot.WorkloadIdentitySelector{
 					Name: "my-workload-identity",
 				},
 				CredentialLifetime: bot.CredentialLifetime{
@@ -57,7 +57,7 @@ func TestWorkloadIdentityAPIService_YAML(t *testing.T) {
 			name: "minimal",
 			in: WorkloadIdentityAPIService{
 				Listen: "tcp://0.0.0.0:4040",
-				Selector: WorkloadIdentitySelector{
+				Selector: bot.WorkloadIdentitySelector{
 					Name: "my-workload-identity",
 				},
 			},
@@ -74,14 +74,14 @@ func TestWorkloadIdentityAPIService_CheckAndSetDefaults(t *testing.T) {
 			name: "valid",
 			in: func() *WorkloadIdentityAPIService {
 				return &WorkloadIdentityAPIService{
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Name: "my-workload-identity",
 					},
 					Listen: "tcp://0.0.0.0:4040",
 				}
 			},
 			want: &WorkloadIdentityAPIService{
-				Selector: WorkloadIdentitySelector{
+				Selector: bot.WorkloadIdentitySelector{
 					Name: "my-workload-identity",
 				},
 				Listen: "tcp://0.0.0.0:4040",
@@ -96,7 +96,7 @@ func TestWorkloadIdentityAPIService_CheckAndSetDefaults(t *testing.T) {
 			name: "valid with labels",
 			in: func() *WorkloadIdentityAPIService {
 				return &WorkloadIdentityAPIService{
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Labels: map[string][]string{
 							"key": {"value"},
 						},
@@ -105,7 +105,7 @@ func TestWorkloadIdentityAPIService_CheckAndSetDefaults(t *testing.T) {
 				}
 			},
 			want: &WorkloadIdentityAPIService{
-				Selector: WorkloadIdentitySelector{
+				Selector: bot.WorkloadIdentitySelector{
 					Labels: map[string][]string{
 						"key": {"value"},
 					},
@@ -122,7 +122,7 @@ func TestWorkloadIdentityAPIService_CheckAndSetDefaults(t *testing.T) {
 			name: "missing selectors",
 			in: func() *WorkloadIdentityAPIService {
 				return &WorkloadIdentityAPIService{
-					Selector: WorkloadIdentitySelector{},
+					Selector: bot.WorkloadIdentitySelector{},
 					Listen:   "tcp://0.0.0.0:4040",
 				}
 			},
@@ -132,7 +132,7 @@ func TestWorkloadIdentityAPIService_CheckAndSetDefaults(t *testing.T) {
 			name: "too many selectors",
 			in: func() *WorkloadIdentityAPIService {
 				return &WorkloadIdentityAPIService{
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Name: "my-workload-identity",
 						Labels: map[string][]string{
 							"key": {"value"},
@@ -147,7 +147,7 @@ func TestWorkloadIdentityAPIService_CheckAndSetDefaults(t *testing.T) {
 			name: "missing listen",
 			in: func() *WorkloadIdentityAPIService {
 				return &WorkloadIdentityAPIService{
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Name: "my-workload-identity",
 					},
 				}
