+++ lib/tbot/config/service_workload_identity_aws_ra_test.go
@@ -20,6 +20,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/gravitational/teleport/lib/tbot/bot"
 	"github.com/gravitational/teleport/lib/tbot/bot/destination"
 	"github.com/gravitational/teleport/lib/tbot/botfs"
 )
@@ -33,7 +34,7 @@ func TestWorkloadIdentityAWSRAService_YAML(t *testing.T) {
 			name: "full",
 			in: WorkloadIdentityAWSRAService{
 				Destination: dest,
-				Selector: WorkloadIdentitySelector{
+				Selector: bot.WorkloadIdentitySelector{
 					Name: "my-workload-identity",
 				},
 				SessionDuration:         time.Minute * 59,
@@ -51,7 +52,7 @@ func TestWorkloadIdentityAWSRAService_YAML(t *testing.T) {
 			name: "simple",
 			in: WorkloadIdentityAWSRAService{
 				Destination: dest,
-				Selector: WorkloadIdentitySelector{
+				Selector: bot.WorkloadIdentitySelector{
 					Name: "my-workload-identity",
 				},
 				SessionDuration:        time.Minute * 59,
@@ -74,7 +75,7 @@ func TestWorkloadIdentityAWSRAService_CheckAndSetDefaults(t *testing.T) {
 			name: "valid",
 			in: func() *WorkloadIdentityAWSRAService {
 				return &WorkloadIdentityAWSRAService{
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Name: "my-workload-identity",
 					},
 					Destination: &destination.Directory{
@@ -89,7 +90,7 @@ func TestWorkloadIdentityAWSRAService_CheckAndSetDefaults(t *testing.T) {
 				}
 			},
 			want: &WorkloadIdentityAWSRAService{
-				Selector: WorkloadIdentitySelector{
+				Selector: bot.WorkloadIdentitySelector{
 					Name: "my-workload-identity",
 				},
 				Destination: &destination.Directory{
@@ -109,7 +110,7 @@ func TestWorkloadIdentityAWSRAService_CheckAndSetDefaults(t *testing.T) {
 			name: "valid with labels",
 			in: func() *WorkloadIdentityAWSRAService {
 				return &WorkloadIdentityAWSRAService{
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Labels: map[string][]string{
 							"key": {"value"},
 						},
@@ -132,7 +133,7 @@ func TestWorkloadIdentityAWSRAService_CheckAndSetDefaults(t *testing.T) {
 			name: "missing selectors",
 			in: func() *WorkloadIdentityAWSRAService {
 				return &WorkloadIdentityAWSRAService{
-					Selector: WorkloadIdentitySelector{},
+					Selector: bot.WorkloadIdentitySelector{},
 					Destination: &destination.Directory{
 						Path:     "/opt/machine-id",
 						ACLs:     botfs.ACLOff,
@@ -150,7 +151,7 @@ func TestWorkloadIdentityAWSRAService_CheckAndSetDefaults(t *testing.T) {
 			name: "too many selectors",
 			in: func() *WorkloadIdentityAWSRAService {
 				return &WorkloadIdentityAWSRAService{
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Name: "my-workload-identity",
 						Labels: map[string][]string{
 							"key": {"value"},
@@ -174,7 +175,7 @@ func TestWorkloadIdentityAWSRAService_CheckAndSetDefaults(t *testing.T) {
 			in: func() *WorkloadIdentityAWSRAService {
 				return &WorkloadIdentityAWSRAService{
 					Destination: nil,
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Name: "my-workload-identity",
 					},
 					RoleARN:        "arn:aws:iam::123456789012:role/example-role",
@@ -189,7 +190,7 @@ func TestWorkloadIdentityAWSRAService_CheckAndSetDefaults(t *testing.T) {
 			name: "missing role_arn",
 			in: func() *WorkloadIdentityAWSRAService {
 				return &WorkloadIdentityAWSRAService{
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Name: "my-workload-identity",
 					},
 					Destination: &destination.Directory{
@@ -208,7 +209,7 @@ func TestWorkloadIdentityAWSRAService_CheckAndSetDefaults(t *testing.T) {
 			name: "missing trust_anchor_arn",
 			in: func() *WorkloadIdentityAWSRAService {
 				return &WorkloadIdentityAWSRAService{
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Name: "my-workload-identity",
 					},
 					Destination: &destination.Directory{
@@ -227,7 +228,7 @@ func TestWorkloadIdentityAWSRAService_CheckAndSetDefaults(t *testing.T) {
 			name: "missing profile_arn",
 			in: func() *WorkloadIdentityAWSRAService {
 				return &WorkloadIdentityAWSRAService{
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Name: "my-workload-identity",
 					},
 					Destination: &destination.Directory{
@@ -246,7 +247,7 @@ func TestWorkloadIdentityAWSRAService_CheckAndSetDefaults(t *testing.T) {
 			name: "invalid role arn",
 			in: func() *WorkloadIdentityAWSRAService {
 				return &WorkloadIdentityAWSRAService{
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Name: "my-workload-identity",
 					},
 					Destination: &destination.Directory{
@@ -266,7 +267,7 @@ func TestWorkloadIdentityAWSRAService_CheckAndSetDefaults(t *testing.T) {
 			name: "invalid trust anchor arn",
 			in: func() *WorkloadIdentityAWSRAService {
 				return &WorkloadIdentityAWSRAService{
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Name: "my-workload-identity",
 					},
 					Destination: &destination.Directory{
@@ -286,7 +287,7 @@ func TestWorkloadIdentityAWSRAService_CheckAndSetDefaults(t *testing.T) {
 			name: "invalid profile arn",
 			in: func() *WorkloadIdentityAWSRAService {
 				return &WorkloadIdentityAWSRAService{
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Name: "my-workload-identity",
 					},
 					Destination: &destination.Directory{
@@ -306,7 +307,7 @@ func TestWorkloadIdentityAWSRAService_CheckAndSetDefaults(t *testing.T) {
 			name: "invalid region",
 			in: func() *WorkloadIdentityAWSRAService {
 				return &WorkloadIdentityAWSRAService{
-					Selector: WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Name: "my-workload-identity",
 					},
 					Destination: &destination.Directory{
