+++ lib/tbot/config/service_workload_identity_jwt.go
@@ -40,7 +40,7 @@ type WorkloadIdentityJWTService struct {
 	Name string `yaml:"name,omitempty"`
 	// Selector is the selector for the WorkloadIdentity resource that will be
 	// used to issue WICs.
-	Selector WorkloadIdentitySelector `yaml:"selector"`
+	Selector bot.WorkloadIdentitySelector `yaml:"selector"`
 	// Destination is where the credentials should be written to.
 	Destination destination.Destination `yaml:"destination"`
 	// Audiences is the list of audiences that the JWT should be valid for.
