+++ lib/tbot/config/service_workload_identity_x509.go
@@ -34,34 +34,6 @@ var (
 	_ Initable      = &WorkloadIdentityX509Service{}
 )
 
-// WorkloadIdentitySelector allows the user to select which WorkloadIdentity
-// resource should be used.
-//
-// Only one of Name or Labels can be set.
-type WorkloadIdentitySelector struct {
-	// Name is the name of a specific WorkloadIdentity resource.
-	Name string `yaml:"name"`
-	// Labels is a set of labels that the WorkloadIdentity resource must have.
-	Labels map[string][]string `yaml:"labels,omitempty"`
-}
-
-// CheckAndSetDefaults checks the WorkloadIdentitySelector values and sets any
-// defaults.
-func (s *WorkloadIdentitySelector) CheckAndSetDefaults() error {
-	switch {
-	case s.Name == "" && len(s.Labels) == 0:
-		return trace.BadParameter("one of ['name', 'labels'] must be set")
-	case s.Name != "" && len(s.Labels) > 0:
-		return trace.BadParameter("at most one of ['name', 'labels'] can be set")
-	}
-	for k, v := range s.Labels {
-		if len(v) == 0 {
-			return trace.BadParameter("labels[%s]: must have at least one value", k)
-		}
-	}
-	return nil
-}
-
 // WorkloadIdentityX509Service is the configuration for the WorkloadIdentityX509Service
 // Emulates the output of https://github.com/spiffe/spiffe-helper
 type WorkloadIdentityX509Service struct {
@@ -69,7 +41,7 @@ type WorkloadIdentityX509Service struct {
 	Name string `yaml:"name,omitempty"`
 	// Selector is the selector for the WorkloadIdentity resource that will be
 	// used to issue WICs.
-	Selector WorkloadIdentitySelector `yaml:"selector"`
+	Selector bot.WorkloadIdentitySelector `yaml:"selector"`
 	// Destination is where the credentials should be written to.
 	Destination destination.Destination `yaml:"destination"`
 	// IncludeFederatedTrustBundles controls whether to include federated trust
