+++ lib/tbot/service_workload_identity_api_test.go
@@ -39,6 +39,7 @@ import (
 	workloadidentityv1pb "github.com/gravitational/teleport/api/gen/proto/go/teleport/workloadidentity/v1"
 	"github.com/gravitational/teleport/api/types"
 	apiutils "github.com/gravitational/teleport/api/utils"
+	"github.com/gravitational/teleport/lib/tbot/bot"
 	"github.com/gravitational/teleport/lib/tbot/config"
 	"github.com/gravitational/teleport/lib/utils/log/logtest"
 	"github.com/gravitational/teleport/tool/teleport/testenv"
@@ -100,7 +101,7 @@ func TestBotWorkloadIdentityAPI(t *testing.T) {
 	onboarding, _ := makeBot(t, rootClient, "api", role.GetName())
 	botConfig := defaultBotConfig(t, process, onboarding, config.ServiceConfigs{
 		&config.WorkloadIdentityAPIService{
-			Selector: config.WorkloadIdentitySelector{
+			Selector: bot.WorkloadIdentitySelector{
 				Name: workloadIdentity.GetMetadata().GetName(),
 			},
 			Listen: listenAddr.String(),
