+++ lib/tbot/service_workload_identity_aws_ra_test.go
@@ -35,6 +35,7 @@ import (
 	workloadidentityv1pb "github.com/gravitational/teleport/api/gen/proto/go/teleport/workloadidentity/v1"
 	"github.com/gravitational/teleport/api/types"
 	apiutils "github.com/gravitational/teleport/api/utils"
+	"github.com/gravitational/teleport/lib/tbot/bot"
 	"github.com/gravitational/teleport/lib/tbot/bot/destination"
 	"github.com/gravitational/teleport/lib/tbot/config"
 	"github.com/gravitational/teleport/lib/tlsca"
@@ -302,7 +303,7 @@ func TestBotWorkloadIdentityAWSRA(t *testing.T) {
 			onboarding, _ := makeBot(t, rootClient, "ra-test", role.GetName())
 			botConfig := defaultBotConfig(t, process, onboarding, config.ServiceConfigs{
 				&config.WorkloadIdentityAWSRAService{
-					Selector: config.WorkloadIdentitySelector{
+					Selector: bot.WorkloadIdentitySelector{
 						Name: workloadIdentity.GetMetadata().GetName(),
 					},
 					Destination: &destination.Directory{
