+++ lib/tbot/service_workload_identity_jwt_test.go
@@ -30,6 +30,7 @@ import (
 	workloadidentityv1pb "github.com/gravitational/teleport/api/gen/proto/go/teleport/workloadidentity/v1"
 	"github.com/gravitational/teleport/api/types"
 	apiutils "github.com/gravitational/teleport/api/utils"
+	"github.com/gravitational/teleport/lib/tbot/bot"
 	"github.com/gravitational/teleport/lib/tbot/bot/destination"
 	"github.com/gravitational/teleport/lib/tbot/config"
 	"github.com/gravitational/teleport/lib/utils/log/logtest"
@@ -88,7 +89,7 @@ func TestBotWorkloadIdentityJWT(t *testing.T) {
 		onboarding, _ := makeBot(t, rootClient, "by-name", role.GetName())
 		botConfig := defaultBotConfig(t, process, onboarding, config.ServiceConfigs{
 			&config.WorkloadIdentityJWTService{
-				Selector: config.WorkloadIdentitySelector{
+				Selector: bot.WorkloadIdentitySelector{
 					Name: workloadIdentity.GetMetadata().GetName(),
 				},
 				Destination: &destination.Directory{
@@ -119,7 +120,7 @@ func TestBotWorkloadIdentityJWT(t *testing.T) {
 		onboarding, _ := makeBot(t, rootClient, "by-labels", role.GetName())
 		botConfig := defaultBotConfig(t, process, onboarding, config.ServiceConfigs{
 			&config.WorkloadIdentityJWTService{
-				Selector: config.WorkloadIdentitySelector{
+				Selector: bot.WorkloadIdentitySelector{
 					Labels: map[string][]string{
 						"foo": {"bar"},
 					},
