+++ lib/tbot/service_workload_identity_x509_test.go
@@ -37,6 +37,7 @@ import (
 	apiutils "github.com/gravitational/teleport/api/utils"
 	"github.com/gravitational/teleport/lib/service"
 	"github.com/gravitational/teleport/lib/services"
+	"github.com/gravitational/teleport/lib/tbot/bot"
 	"github.com/gravitational/teleport/lib/tbot/bot/destination"
 	"github.com/gravitational/teleport/lib/tbot/config"
 	"github.com/gravitational/teleport/lib/tlsca"
@@ -106,7 +107,7 @@ func TestBotWorkloadIdentityX509(t *testing.T) {
 		onboarding, _ := makeBot(t, rootClient, "by-name", role.GetName())
 		botConfig := defaultBotConfig(t, process, onboarding, config.ServiceConfigs{
 			&config.WorkloadIdentityX509Service{
-				Selector: config.WorkloadIdentitySelector{
+				Selector: bot.WorkloadIdentitySelector{
 					Name: workloadIdentity.GetMetadata().GetName(),
 				},
 				Destination: &destination.Directory{
@@ -150,7 +151,7 @@ func TestBotWorkloadIdentityX509(t *testing.T) {
 		onboarding, _ := makeBot(t, rootClient, "by-labels", role.GetName())
 		botConfig := defaultBotConfig(t, process, onboarding, config.ServiceConfigs{
 			&config.WorkloadIdentityX509Service{
-				Selector: config.WorkloadIdentitySelector{
+				Selector: bot.WorkloadIdentitySelector{
 					Labels: map[string][]string{
 						"foo": {"bar"},
 					},
