+++ lib/tbot/workloadidentity/issue.go
@@ -29,7 +29,7 @@ import (
 	apiclient "github.com/gravitational/teleport/api/client"
 	workloadidentityv1pb "github.com/gravitational/teleport/api/gen/proto/go/teleport/workloadidentity/v1"
 	"github.com/gravitational/teleport/lib/cryptosuites"
-	"github.com/gravitational/teleport/lib/tbot/config"
+	"github.com/gravitational/teleport/lib/tbot/bot"
 	"github.com/gravitational/teleport/lib/tbot/workloadidentity/attrs"
 )
 
@@ -81,7 +81,7 @@ func IssueX509WorkloadIdentity(
 	ctx context.Context,
 	log *slog.Logger,
 	clt authClient,
-	workloadIdentity config.WorkloadIdentitySelector,
+	workloadIdentity bot.WorkloadIdentitySelector,
 	ttl time.Duration,
 	attest *attrs.WorkloadAttrs,
 ) ([]*workloadidentityv1pb.Credential, crypto.Signer, error) {
@@ -183,7 +183,7 @@ func IssueJWTWorkloadIdentity(
 	ctx context.Context,
 	log *slog.Logger,
 	clt *apiclient.Client,
-	workloadIdentity config.WorkloadIdentitySelector,
+	workloadIdentity bot.WorkloadIdentitySelector,
 	audiences []string,
 	ttl time.Duration,
 	attest *attrs.WorkloadAttrs,
